ui disable rec and play during playback, update meters to better calculation
