+++ cli.c
@@ -1,0 +1,73 @@
+#include "audio.h"
+
+// Function to check if a key was pressed
+bool keyPressed()
+{
+  struct timeval tv = {0L, 0L};
+  fd_set fds;
+  FD_ZERO(&fds);
+  FD_SET(0, &fds); // STDIN_FILENO is 0
+  return select(1, &fds, NULL, NULL, &tv) > 0;
+}
+
+// Function to get the pressed key without waiting for Enter key
+int getCharNonBlocking()
+{
+  struct termios oldt, newt;
+  int ch;
+  tcgetattr(STDIN_FILENO, &oldt); // Get current terminal attributes
+  newt = oldt;
+  newt.c_lflag &= ~(ICANON | ECHO);        // Disable buffering and echo
+  tcsetattr(STDIN_FILENO, TCSANOW, &newt); // Apply new settings
+
+  if (keyPressed())
+  { // Check if a key was pressed
+    ch = getchar();
+  }
+  else
+  {
+    ch = -1; // No key was pressed
+  }
+
+  tcsetattr(STDIN_FILENO, TCSANOW, &oldt); // Restore terminal attributes
+  return ch;
+}
+
+void inputStartTime()
+{
+  float time;
+  printf("Enter new start time in seconds (current: %f): ", startTimeInSeconds);
+  if (scanf("%f", &time) == 1)
+  { // Ensure scanf successfully reads a float
+    // Validate the input time range
+    if (time >= 0 && time <= 3600)
+    { // 0 to 3600 seconds range
+      updateStartTime(time);
+      printf("Start time updated to %f seconds.\n", time);
+    }
+    else
+    {
+      printf("Invalid input: start time must be between 0 and 3600 seconds.\n");
+    }
+  }
+  else
+  {
+    // Clear the input buffer to prevent infinite loops in case of invalid input
+    while (getchar() != '\n')
+      ;
+    printf("Invalid input: please enter a numeric value.\n");
+  }
+}
+
+// TESTING
+// int main() {
+//   initAudio();
+//   onStartRecording();
+//   sleep(3);
+//   onStopRecording();
+//   onStartPlaying();
+//   sleep(3);
+//   onStopPlaying();
+
+//   cleanupAudio();
+// }
